+++ aula-01/java/com/github/jc00973/cs/aula1/ex16.java
@@ -1,0 +1,20 @@
+package com.github.jc00973.cs.aula1;
+
+public class ex16 {
+	
+	public static double mdc(double a, double b) {
+		
+		while(b!=0) {
+			double m = a%b;
+			a = b;
+			b = m;
+		}
+		
+		return a;
+	}
+	
+	public static void main(String args[]) {
+		System.out.println(mdc(10, 5));
+	}
+
+}
