+++ aula-01/java/com/github/jc00973/cs/aula1/ex18.java
@@ -1,0 +1,17 @@
+package com.github.jc00973.cs.aula1;
+
+public class ex18 {
+	
+	public static double horner(int x, int g, int a[]) {
+		
+		double p = a[g];
+		int i = g - 1;
+		
+		while(0<=i) {
+			p = p * x + a[i];
+			i = i - 1;
+		}
+		
+		return p;
+	}
+}
